backend/requirements.txt
asgiref==3.9.2
Django==5.2.7
sqlparse==0.5.3
djangorestframework
django-cors-headers
djangorestframework-simplejwt
channels
uvicorn
uvicorn[standard]
gunicorn
whitenoise
pillow